Projeto 285 - Nova Conta Online
git-tfs-id: [http://alm.cecred.coop.br:8080/tfs/DefaultCollection]$/FONTES/2018_02;C17063

+++ oracle/CECRED/packages/folh0002.pck
@@ -373,6 +373,7 @@ CREATE OR REPLACE PACKAGE CECRED.FOLH0002 AS
                                   ,pr_gravarpg   IN NUMBER
                                   ,pr_dsdirarq   IN VARCHAR2
                                   ,pr_dsarquiv   IN VARCHAR2
+                                  ,pr_dsdspscp   IN NUMBER
                                   ,pr_cdcritic   OUT PLS_INTEGER
                                   ,pr_dscritic   OUT VARCHAR2);
 
@@ -428,7 +429,10 @@ CREATE OR REPLACE PACKAGE CECRED.FOLH0002 AS
                                  ,pr_dtfimsel  IN DATE
                                  ,pr_insituac  IN INTEGER
                                  ,pr_tpemissa  IN VARCHAR2
+                                 ,pr_iddspscp  IN NUMBER
                                  ,pr_nmarquiv OUT VARCHAR2
+                                 ,pr_dssrvarq OUT VARCHAR2
+                                 ,pr_dsdirarq OUT VARCHAR2
                                  ,pr_dscritic OUT VARCHAR2);
 
   --Procedure para impimir os comprovantes  HOLERITE
@@ -436,6 +440,7 @@ CREATE OR REPLACE PACKAGE CECRED.FOLH0002 AS
                                     ,pr_nrdconta IN NUMBER
                                     ,pr_idtipfol IN NUMBER
                                     ,pr_rowidpfp IN VARCHAR2
+                                    ,pr_iddspscp IN NUMBER
                                     ,pr_retxml   OUT VARCHAR2
                                     ,pr_cdcritic OUT PLS_INTEGER
                                     ,pr_dscritic OUT VARCHAR2);
@@ -5724,9 +5729,9 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
                                       ,pfp.flsitdeb
                                       ,pfp.flsitcre
                                       ,pfp.dsobscre     
-               ,NVL(to_char(pfp.dthordeb,'dd/mm/RRRR hh24:mi'),to_char(pfp.dtdebito,'dd/mm/RRRR')) dthordeb
+               ,NVL(to_char(pfp.dthordeb,'dd/mm/rr hh24:mi'),to_char(pfp.dtdebito,'dd/mm/rr')) dthordeb
                ,DECODE(pfp.flsitdeb,0,NVL(TRIM(pfp.dsobsdeb),'Pagamento aguardando data do débito.'),'Débito efetuado com sucesso') dsobservdeb
-               ,NVL(to_char(pfp.dthorcre,'dd/mm/RRRR hh24:mi'),to_char(pfp.dtcredit,'dd/mm/RRRR')) dthorcre
+               ,NVL(to_char(pfp.dthorcre,'dd/mm/rr hh24:mi'),to_char(pfp.dtcredit,'dd/mm/rr')) dthorcre
                                       ,DECODE(pfp.flsitcre,0,nvl(TRIM(pfp.dsobscre),'Pagamento aguardando data do crédito.'),DECODE((SELECT COUNT(1)
                                                                        FROM craplfp lfp
                                                                       WHERE lfp.cdcooper = pfp.cdcooper
@@ -6044,13 +6049,13 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
           
           CASE 
             WHEN NVL(rw_registros.dthordeb,'') = '' THEN vr_tab_pagamento(vr_idx_pagto).dthrdebi := ''; 
-            WHEN LENGTH(rw_registros.dthordeb) = 16 THEN vr_tab_pagamento(vr_idx_pagto).dthrdebi := rw_registros.dthordeb;
+            WHEN LENGTH(rw_registros.dthordeb) = 14 THEN vr_tab_pagamento(vr_idx_pagto).dthrdebi := rw_registros.dthordeb;
             ELSE vr_tab_pagamento(vr_idx_pagto).dthrdebi := rw_registros.dthordeb || ' 00:00';
           END CASE;
 
           CASE 
             WHEN NVL(rw_registros.dthorcre,'') = '' THEN vr_tab_pagamento(vr_idx_pagto).dthrcred := '';
-            WHEN LENGTH(rw_registros.dthorcre) = 16 THEN vr_tab_pagamento(vr_idx_pagto).dthrcred := rw_registros.dthorcre;
+            WHEN LENGTH(rw_registros.dthorcre) = 14 THEN vr_tab_pagamento(vr_idx_pagto).dthrcred := rw_registros.dthorcre;
             ELSE vr_tab_pagamento(vr_idx_pagto).dthrcred := rw_registros.dthorcre || ' 00:00';
           END CASE;                    
 
@@ -8122,6 +8127,7 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
                                   ,pr_gravarpg   IN NUMBER
                                   ,pr_dsdirarq   IN VARCHAR2
                                   ,pr_dsarquiv   IN VARCHAR2
+                                  ,pr_dsdspscp   IN NUMBER
                                   ,pr_cdcritic   OUT PLS_INTEGER
                                   ,pr_dscritic   OUT VARCHAR2) IS
    ---------------------------------------------------------------------------------------------------------------
@@ -8143,6 +8149,12 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
    --             27/01/2016 - Incluir controle de lançamentos sem crédito (Marcos-Supero)
    ---------------------------------------------------------------------------------------------------------------
 
+      CURSOR cr_crapcop(pr_cdcooper crapcop.cdcooper%TYPE) IS
+        SELECT cop.dsdircop
+          FROM crapcop cop
+         WHERE cop.cdcooper = pr_cdcooper;
+      rw_crapcop cr_crapcop%ROWTYPE;  
+    
       -- Busca empresa cadastrada pelo usuario conectado
       CURSOR cr_crapemp(p_cdcooper crapcop.cdcooper%TYPE
                        ,p_nrdconta crapemp.nrdconta%TYPE) IS
@@ -8298,22 +8310,44 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
       vr_indvalid := 'N';
       vr_tab_pgto.DELETE;
       vr_tab_origem.DELETE;
+      
+      --> Verificar cooperativa
+      OPEN cr_crapcop (pr_cdcooper => pr_cdcooper);    
+      FETCH cr_crapcop INTO rw_crapcop;
+        
+      --> verificar se encontra registro
+      IF cr_crapcop%NOTFOUND THEN      
+        CLOSE cr_crapcop;    
+          
+        pr_dscritic := 'Cooperativa de destino nao cadastrada.';      
+        RAISE vr_erro;      
+      ELSE
+        CLOSE cr_crapcop;
+      END IF;       
 
-      -- Busca o diretório do upload do arquivo
-      vr_dsdireto := GENE0001.fn_diretorio(pr_tpdireto => 'C'
-                                          ,pr_cdcooper => pr_cdcooper
-                                          ,pr_nmsubdir => 'upload');
+      IF pr_dsdspscp = 0 THEN -- Diretorio de upload do gnusites
+        -- Busca o diretório do upload do arquivo
+        vr_dsdireto := GENE0001.fn_diretorio(pr_tpdireto => 'C'
+                                            ,pr_cdcooper => pr_cdcooper
+                                            ,pr_nmsubdir => 'upload');
 
-      -- Realizar a cópia do arquivo
-      GENE0001.pc_OScommand_Shell(gene0001.fn_param_sistema('CRED',0,'SCRIPT_RECEBE_ARQUIVOS')||pr_dsdirarq||pr_dsarquiv||' S'
-                                 ,pr_typ_saida   => vr_typ_said
-                                 ,pr_des_saida   => vr_des_erro);
-      -- Testar erro
-      IF vr_typ_said = 'ERR' THEN
-        -- O comando shell executou com erro, gerar log e sair do processo
-        pr_dscritic := 'Erro no arquivo: ' || vr_des_erro;
-        RAISE vr_erro;
-      END IF;
+        -- Realizar a cópia do arquivo
+        GENE0001.pc_OScommand_Shell(gene0001.fn_param_sistema('CRED',0,'SCRIPT_RECEBE_ARQUIVOS')||pr_dsdirarq||pr_dsarquiv||' S'
+                                   ,pr_typ_saida   => vr_typ_said
+                                   ,pr_des_saida   => vr_des_erro);
+        -- Testar erro
+        IF vr_typ_said = 'ERR' THEN
+          -- O comando shell executou com erro, gerar log e sair do processo
+          pr_dscritic := 'Erro no arquivo: ' || vr_des_erro;
+          RAISE vr_erro;
+        END IF;
+      ELSE
+        vr_dsdireto := gene0001.fn_diretorio('C',0)                                ||
+                       gene0001.fn_param_sistema('CRED',0,'PATH_DOWNLOAD_ARQUIVO') ||
+                       '/'                                                         ||
+                       rw_crapcop.dsdircop                                         ||
+                       '/upload';
+      END IF;        
 
       -- Se não existir o arquivo
       IF NOT gene0001.fn_exis_arquivo(pr_caminho => vr_dsdireto||'/'||pr_dsarquiv) THEN
@@ -9004,9 +9038,13 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
                 pfp.dtcredit,
                 pfp.dtdebito,
                 pfp.qtregpag,
-                pfp.vllctpag
-           FROM crappfp pfp
-          WHERE pfp.rowid = pr_rowidpfp;
+                pfp.vllctpag,
+                pfp.flsitcre,
+                dat.dtmvtolt - TRUNC(pfp.dthorcre) qtsubtra
+           FROM crappfp pfp,
+                crapdat dat
+          WHERE pfp.rowid = pr_rowidpfp 
+            AND dat.cdcooper = pfp.cdcooper;
       rw_crappfp cr_crappfp%ROWTYPE;
 
 
@@ -9017,12 +9055,13 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
       vr_msgtela VARCHAR2(32726) := '';
       vr_cdcritic INTEGER := 0;
       vr_dscritic VARCHAR2(4000) := '';
+      vr_qtdiaenv NUMBER(3);
    BEGIN
 
       -- Busca as informacoes da empresa atraves da conta
       OPEN cr_crappfp(pr_rowidpfp);
       FETCH cr_crappfp INTO rw_crappfp;
-      CLOSE cr_crappfp;
+      CLOSE cr_crappfp;      
 
       OPEN cr_craplfp(rw_crappfp.cdcooper
                      ,rw_crappfp.cdempres
@@ -9036,6 +9075,21 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
          ELSIF vr_existe < rw_crappfp.qtregpag THEN
               vr_msgtela := 'Pagamento com (' || vr_existe || ') comprovante(s) carregado(s). Carregue novos ou atualize.';
          END IF;
+      END IF;
+      
+      -- Busca a Qtde dias para envio comprovantes
+      vr_qtdiaenv := GENE0001.fn_param_sistema(pr_nmsistem => 'CRED'
+                                              ,pr_cdcooper => rw_crappfp.cdcooper
+                                              ,pr_cdacesso => 'FOLHAIB_QTD_DIA_ENV_COMP');
+                                              
+      IF rw_crappfp.qtsubtra > vr_qtdiaenv THEN        
+        vr_cdcritic := 'O envio dos comprovantes deve ser efetuado no máximo ' || TO_CHAR(vr_qtdiaenv) || ' dia(s) após o pagamento.';
+        RAISE vr_erro;
+      END IF;     
+      
+      IF rw_crappfp.flsitcre NOT IN (1,2) THEN
+        vr_cdcritic := 'Apenas registros creditados podem ser enviados.';
+        RAISE vr_erro;        
       END IF;
 
        -- Monta documento XML
@@ -9491,7 +9545,10 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
                                  ,pr_dtfimsel  IN DATE
                                  ,pr_insituac  IN INTEGER
                                  ,pr_tpemissa  IN VARCHAR2
+                                 ,pr_iddspscp  IN NUMBER
                                  ,pr_nmarquiv OUT VARCHAR2
+                                 ,pr_dssrvarq OUT VARCHAR2
+                                 ,pr_dsdirarq OUT VARCHAR2
                                  ,pr_dscritic OUT VARCHAR2) IS
    ---------------------------------------------------------------------------------------------------------------
    --  Programa : pc_gera_relatorio_ib              Antigo:
@@ -9943,24 +10000,38 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
       dbms_lob.close(vr_des_xml);
       dbms_lob.freetemporary(vr_des_xml);
 
-      -- Copia o PDF para o IB
-      GENE0002.pc_efetua_copia_arq_ib(pr_cdcooper => pr_cdcooper,
-                                      pr_nmarqpdf => vr_nmdireto || '/' || vr_nmarquiv,
-                                      pr_des_erro => vr_des_erro);
-      -- Testar se houve erro
-      IF vr_des_erro IS NOT NULL THEN
-        -- Gerar excecao
-        RAISE vr_exc_erro;
-      END IF;
+      IF pr_iddspscp = 0 THEN
+        -- Copia o PDF para o IB
+        GENE0002.pc_efetua_copia_arq_ib(pr_cdcooper => pr_cdcooper,
+                                        pr_nmarqpdf => vr_nmdireto || '/' || vr_nmarquiv,
+                                        pr_des_erro => vr_des_erro);
+        -- Testar se houve erro
+        IF vr_des_erro IS NOT NULL THEN
+          -- Gerar excecao
+          RAISE vr_exc_erro;
+        END IF;
 
-      -- Remove o arquivo XML fisico de envio
-      GENE0001.pc_OScommand (pr_typ_comando => 'S'
-                            ,pr_des_comando => 'rm '||vr_nmdireto || '/' || vr_nmarquiv||' 2> /dev/null'
-                            ,pr_typ_saida   => vr_des_reto
-                            ,pr_des_saida   => vr_des_erro);
-      -- Se ocorreu erro dar RAISE
-      IF vr_des_reto = 'ERR' THEN
-        RAISE vr_exc_erro;
+        -- Remove o arquivo XML fisico de envio
+        GENE0001.pc_OScommand (pr_typ_comando => 'S'
+                              ,pr_des_comando => 'rm '||vr_nmdireto || '/' || vr_nmarquiv||' 2> /dev/null'
+                              ,pr_typ_saida   => vr_des_reto
+                              ,pr_des_saida   => vr_des_erro);
+        -- Se ocorreu erro dar RAISE
+        IF vr_des_reto = 'ERR' THEN
+          RAISE vr_exc_erro;
+        END IF;
+      ELSE
+        gene0002.pc_copia_arq_para_download(pr_cdcooper => pr_cdcooper
+                                           ,pr_dsdirecp => vr_nmdireto||'/'
+                                           ,pr_nmarqucp => vr_nmarquiv
+                                           ,pr_flgcopia => 0
+                                           ,pr_dssrvarq => pr_dssrvarq
+                                           ,pr_dsdirarq => pr_dsdirarq
+                                           ,pr_des_erro => vr_dscritic);
+          
+        IF vr_dscritic IS NOT NULL AND TRIM(vr_dscritic) <> ' ' THEN
+          RAISE vr_exc_erro;
+        END IF;
       END IF;
 
       -- Seta o retorno
@@ -9986,6 +10057,7 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
                                       ,pr_nrdconta IN NUMBER
                                       ,pr_idtipfol IN NUMBER
                                       ,pr_rowidpfp IN VARCHAR2
+                                      ,pr_iddspscp IN NUMBER
                                       ,pr_retxml   OUT VARCHAR2
                                       ,pr_cdcritic OUT PLS_INTEGER
                                       ,pr_dscritic OUT VARCHAR2) IS
@@ -10087,6 +10159,8 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
       vr_des_reto VARCHAR2(3);
       vr_nmextemp VARCHAR2(100);
       vr_dsdpagto VARCHAR2(4000);
+      vr_dssrvarq VARCHAR2(500);
+      vr_dsdirarq VARCHAR2(500);
 
       vr_dsxmlrel    CLOB;
       vr_dsdtexto    VARCHAR2(32000);
@@ -10337,30 +10411,46 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
 
           IF pr_dscritic IS NULL THEN
 
-              -- Copia o PDF para o IB
-              GENE0002.pc_efetua_copia_arq_ib(pr_cdcooper => pr_cdcooper,
-                                              pr_nmarqpdf => vr_nmdireto || '/' || vr_nmarquiv,
-                                              pr_des_erro => vr_des_erro);
+              IF pr_iddspscp = 0 THEN
+                -- Copia o PDF para o IB
+                GENE0002.pc_efetua_copia_arq_ib(pr_cdcooper => pr_cdcooper,
+                                                pr_nmarqpdf => vr_nmdireto || '/' || vr_nmarquiv,
+                                                pr_des_erro => vr_des_erro);
 
-             -- Testar se houve erro
-              IF vr_des_erro IS NOT NULL THEN
-                -- Gerar excecao
-                 pr_dscritic := vr_des_erro;
-              END IF;
+               -- Testar se houve erro
+                IF vr_des_erro IS NOT NULL THEN
+                  -- Gerar excecao
+                   pr_dscritic := vr_des_erro;
+                END IF;
 
-              -- Remove o arquivo XML fisico de envio
-              GENE0001.pc_OScommand (pr_typ_comando => 'S'
-                                    ,pr_des_comando => 'rm '||vr_nmdireto || '/' || vr_nmarquiv||' 2> /dev/null'
-                                    ,pr_typ_saida   => vr_des_reto
-                                    ,pr_des_saida   => vr_des_erro);
-              -- Se ocorreu erro dar RAISE
-              IF vr_des_reto = 'ERR' THEN
-                pr_dscritic := vr_des_erro;
+                -- Remove o arquivo XML fisico de envio
+                GENE0001.pc_OScommand (pr_typ_comando => 'S'
+                                      ,pr_des_comando => 'rm '||vr_nmdireto || '/' || vr_nmarquiv||' 2> /dev/null'
+                                      ,pr_typ_saida   => vr_des_reto
+                                      ,pr_des_saida   => vr_des_erro);
+                -- Se ocorreu erro dar RAISE
+                IF vr_des_reto = 'ERR' THEN
+                  pr_dscritic := vr_des_erro;
+                END IF;
+                
+                pr_retxml := '<nmarquiv>' || vr_nmarquiv || '</nmarquiv>';
+              ELSE
+                gene0002.pc_copia_arq_para_download(pr_cdcooper => pr_cdcooper
+                                                   ,pr_dsdirecp => vr_nmdireto||'/'
+                                                   ,pr_nmarqucp => vr_nmarquiv
+                                                   ,pr_flgcopia => 0
+                                                   ,pr_dssrvarq => vr_dssrvarq
+                                                   ,pr_dsdirarq => vr_dsdirarq
+                                                   ,pr_des_erro => pr_dscritic);  
+                                                   
+                pr_retxml := '<nmarquiv>' || vr_nmarquiv          || '</nmarquiv>' ||
+                             '<dssrvarq>' || NVL(vr_dssrvarq,' ') || '</dssrvarq>' ||
+                             '<dsdirarq>' || NVL(vr_dsdirarq,' ') || '</dsdirarq>';              
               END IF;
+              
               COMMIT;
               -- Seta o retorno
               pr_dscritic := NULL;
-              pr_retxml :=  vr_nmarquiv;
           END IF;
    EXCEPTION
      WHEN OTHERS THEN
@@ -10687,6 +10777,7 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
               END flgsegmt
             ,LPAD(gene0002.fn_mask(lfp.nrseqpag,'999999999') || gene0002.fn_mask(lfp.nrseqlfp,'999999999'),25,' ') dsprotoc
             ,pfp.vllctpag
+            ,pfp.idsitapr
         FROM crappfp pfp
         JOIN craplfp lfp 
           ON lfp.cdcooper = pfp.cdcooper
@@ -10839,6 +10930,18 @@ CREATE OR REPLACE PACKAGE BODY CECRED.FOLH0002 AS
         ELSE
           CLOSE cr_crapenc;
         END IF;
+        
+        IF TRUNC(rw_crappfp.dtcredit) = TRUNC(rw_crapdat.dtmvtocd) THEN
+          vr_cdcritic := 0;
+          vr_dscritic := 'Arquivo de retorno dos créditos disponível no próximo dia útil.';
+          RAISE vr_exc_erro;          
+        END IF;
+        
+        IF rw_crappfp.idsitapr = 1 THEN
+          vr_cdcritic := 0;
+          vr_dscritic := 'Necessário aprovação do lote para disponibilização de arquivo de retorno.';
+          RAISE vr_exc_erro;          
+        END IF;        
         
         vr_conttrai := vr_conttrai + 1;
